requirements-3.7.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
cryptography==36.0.1
Django==3.2.11
django-cors-headers==3.11.0
django-extensions==3.1.5
django-filter==21.1
django-rest-authemail==2.1.1
django-rest-knox==4.1.0
djangorestframework==3.13.1
Faker==12.0.0
idna==3.3
psycopg2==2.9.3
pycparser==2.21
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tqdm==4.62.3
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==2.0.2
